tendersaas/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
